extracted utterance builder
