+++ modules/nlu/src/backend/engine2/featurizer2.ts
@@ -1,11 +1,12 @@
-import { MLToolkit } from 'botpress/sdk'
 import _ from 'lodash'
 
-import { Intent, Utterance, UtteranceToken } from './engine2'
-import { computeQuantile } from '../tools/math'
-import { countAlpha, countNum, countSpecial } from '../tools/strings'
 import { MAX_TFIDF, MIN_TFIDF } from '../pipelines/intents/tfidf'
 import { sanitize } from '../pipelines/language/sanitizer'
+import { computeQuantile } from '../tools/math'
+import { countAlpha, countNum, countSpecial } from '../tools/strings'
+
+import { Intent } from './training-pipeline'
+import Utterance, { UtteranceToken } from './utterance'
 
 type FeatureValue = string | number | boolean
